Add LED support

+++ Utils/UberClock.py
@@ -16,6 +16,46 @@ signal_pin_A = 16
 signal_pin_B = 20
 pulse_duration = 0.2 / 1000
 
+# Define the LED
+pwr_led = "/sys/class/leds/PWR"
+act_led = "/sys/class/leds/ACT"
+pwr_current_trigger = None
+act_current_trigger = None
+
+
+def write_to_file(path, value):
+    try:
+        with open(path, 'w') as f:
+            f.write(value + "\n")
+    except OSError as e:
+        print(f"Error writing to {path}: {e}")
+
+
+def save_current_trigger(led):
+    # Save current trigger
+    with open(f"{led}/trigger", 'r') as f:
+        trigger_data = f.read()
+        # Extract the current trigger name, which is enclosed in square brackets
+        start = trigger_data.find('[')
+        end = trigger_data.find(']', start)
+        if start != -1 and end != -1:
+            current_trigger = trigger_data[start + 1:end]  # +1 and end without +1 to exclude the brackets
+        else:
+            current_trigger = None
+            print("No current trigger found.")
+
+        print(current_trigger)  # To verify you're getting the correct current trigger
+    return current_trigger
+    
+
+def turn_led_off(led):
+    write_to_file(f"{led}/brightness", "0")
+
+
+def turn_led_on(led):
+    write_to_file(f"{led}/brightness", "1")
+
+
 def setup_gpio():
     GPIO.setmode(GPIO.BCM)
     GPIO.setup(signal_pin_A, GPIO.OUT)
@@ -42,15 +82,23 @@ def send_signal_at_frequency(frequency):
             next_time += interval
         time.sleep(max(next_time - now, 0))  # Sleep for the remaining time until the next signal
 
+
 def parse_arguments():
     parser = argparse.ArgumentParser(description='Send signals at a set frequency.')
     parser.add_argument('-c', '--frequency', type=float, help='Set the frequency for sending signals.')
     return parser.parse_args()
 
+
 def main():
     args = parse_arguments()
 
     setup_gpio()
+    pwr_current_trigger = save_current_trigger(pwr_led)
+    act_current_trigger = save_current_trigger(act_led)
+    
+    turn_led_off(pwr_led)
+    turn_led_off(act_led)
+
     try:
         if args.frequency:
             send_signal_at_frequency(args.frequency)
@@ -61,45 +109,69 @@ def main():
 
             # Calculate the initial delay to the next minute mark, adjusted for the first signal timing
             next_minute = (base_real_time + timedelta(minutes=1)).replace(second=0, microsecond=0)
-            initial_delay_until_next_signal = (next_minute - base_real_time).total_seconds()
+            initial_delay_until_next_minute_signal = (next_minute - base_real_time).total_seconds()
+
+            # Calculate the initial delay to the next second mark, adjusted for the first signal timing
+            next_second = (base_real_time + timedelta(seconds=1)).replace(microsecond=0)
+            initial_delay_until_next_second_signal = (next_second - base_real_time).total_seconds()
+
 
             # Recalculate if too close to next minute
-            if initial_delay_until_next_signal < 1:
-                next_minute = (base_real_time + timedelta(minutes=2)).replace(second=0, microsecond=0)
+            if initial_delay_until_next_second_signal < 0.1:
+                next_second = (base_real_time + timedelta(seconds=2)).replace(microsecond=0)
+                if initial_delay_until_next_minute_signal < 0.:
+                    next_minute = (base_real_time + timedelta(minutes=2)).replace(second=0, microsecond=0)
+
 
 
             while True:
                 
-                # Calculate the next start and stop signal timings relative to the elapsed time
-                next_start_signal_time = next_minute - timedelta(milliseconds=30.1)  # 30.1 ms before the top of the minute
-                next_stop_signal_time = next_start_signal_time - timedelta(milliseconds=5) # 5 ms before the start signal
+                # Calculate the next timer start and stop signal timings relative to the elapsed time
+                next_timer_start_signal_time = next_minute - timedelta(milliseconds=30.1)  # 30.1 ms before the top of the minute
+                next_timer_stop_signal_time = next_timer_start_signal_time - timedelta(milliseconds=5) # 5 ms before the start signal
                 
+                # Calculate the next LED start and stop signal timings relative to the elapsed time
+                next_led_on_signal_time = next_second - timedelta(milliseconds=0)
+                next_led_off_signal_time = next_led_on_signal_time + timedelta(milliseconds=10)
 
-                # Wait for the next stop signal time
-                print("Waiting to send next stop signal...")
+                if next_timer_stop_signal_time < next_led_on_signal_time:
 
-                while (base_real_time + timedelta(seconds=time.perf_counter() - base_perf_counter)) < next_stop_signal_time:
-                    pass
+                    # Wait for the next stop signal time
+                    print("Waiting to send next stop signal...")
 
-                # Send stop signal. Pulse duration should has short as possible
-                send_signal(signal_pin=signal_pin_B, pulse_duration=pulse_duration)
-                print("\nSignal B!")
+                    while (base_real_time + timedelta(seconds=time.perf_counter() - base_perf_counter)) < next_timer_stop_signal_time:
+                        pass
+
+                    # Send stop signal. Pulse duration should has short as possible
+                    send_signal(signal_pin=signal_pin_B, pulse_duration=pulse_duration)
+                    print("\nSignal B!")
 
 
-                # Wait for the next start signal time
-                print("Waiting to send next start signal...")
+                    # Wait for the next start signal time
+                    print("Waiting to send next start signal...")
 
-                while (base_real_time + timedelta(seconds=time.perf_counter() - base_perf_counter)) < next_start_signal_time:
-                    pass
+                    while (base_real_time + timedelta(seconds=time.perf_counter() - base_perf_counter)) < next_timer_start_signal_time:
+                        pass
 
-                # Send start signal. Pulse duration is not critical 
-                send_signal(signal_pin=signal_pin_A, pulse_duration=10*pulse_duration)
-                print("\nSignal A!")
+                    # Send start signal. Pulse duration is not critical 
+                    send_signal(signal_pin=signal_pin_A, pulse_duration=10*pulse_duration)
+                    print("\nSignal A!")
 
-                # Update the delay for the next cycle, ensuring it accounts for a full minute past the initial delay
-                next_minute += timedelta(minutes=1)
+                    # Update the delay for the next cycle, ensuring it accounts for a full minute past the initial delay
+                    next_minute += timedelta(minutes=1)
+                else:
+                    while (base_real_time + timedelta(seconds=time.perf_counter() - base_perf_counter)) < next_led_on_signal_time:
+                        pass
+                    turn_led_on(pwr_led)
+                    while (base_real_time + timedelta(seconds=time.perf_counter() - base_perf_counter)) < next_led_off_signal_time:
+                        pass
+                    turn_led_off(pwr_led)
+                    next_second += timedelta(seconds=1)
     finally:
         GPIO.cleanup()
+        # Restore the LED trigger
+        write_to_file(f"{pwr_led}/trigger", pwr_current_trigger)
+        write_to_file(f"{act_led}/trigger", act_current_trigger)
 
 
 if __name__ == "__main__":
